+++ src/components/App.jsx
@@ -13,6 +13,7 @@ function App() {
   function updateUrl(newValue) {
     setUrl(newSValue);
   }
+
   return (
     <>
       <WorkSpace />
